Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-docx==1.1.0
python-pptx==0.6.23
pydantic==2.5.2
pydantic-settings==2.1.0
gemini-webapi==1.0.1
httpx==0.27.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
gunicorn==21.2.0
google-generativeai==0.3.2
